requirements.txt
asttokens==2.4.1
Brotli
certifi
charset-normalizer
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.45.1
idna
ipython==8.18.1
jedi==0.19.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
numpy
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
Pillow
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
PySocks
python-dateutil==2.8.2
pytz==2023.3.post1
requests
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.2.0
torch==1.12.1
torchvision
tqdm==4.66.1
traitlets==5.14.0
typing_extensions
tzdata==2023.3
urllib3
utils==1.0.1
transformers
matplotlib
